Members Area Test
This includes a test for the members area based on an idea that Joshua
wanted to show Chaz.

Added session information to all pages.
content\client_area.php is the content file for the client area.
content\staff_area.php is the content file for the staff area.
content\manager_area.php is the content file for the manager area.
includes\navigation.php has been altered to reflect some minor changes
to the inner workings of the code.
includes\navigation_client.php is the sub navigations for the client
area.
members_area.php is the base members area class which was used for this
test.
scripts\members_area_content.php is used to determine which content page
should be loaded based on user access level.
scripts\members_area_pre.php is used to determine if the person
navigating to this page has proper access. Guests will be redirected to
the login page.
scripts\request.php was simply created and left blank.
scripts\view_request.php was simply created and left blank.
template.php will be used for further pages to be created.